refactor: remove PostNavigationClient component
