+++ src/app/(contents)/layout.tsx
@@ -2,7 +2,7 @@ import Footer from "@/components/Footer";
 import ReadingProgress from "@/components/ProgressBar";
 import React, { ReactNode } from "react";
 import MDXToc from "@/components/MDXToc";
-import PostNavigationClient from "@/components/PostNavigationClient";
+import PostNavigation from "@/components/PostNavigation";
 
 interface LayoutProps {
   children: ReactNode;
@@ -16,7 +16,7 @@ export default function PageLayout({ children }: LayoutProps) {
         <div className="hidden md:block md:col-span-2"></div>
         <main className="py-8 px-4 md:col-span-8">
           <article className="mdx-content">{children}</article>
-          <PostNavigationClient />
+          <PostNavigation />
         </main>
         <aside className="hidden md:block md:col-span-2 sticky top-20 h-fit py-8 pr-3">
           <MDXToc />
